backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
langchain>=0.3.13
langchain-groq>=0.2.3
langgraph>=0.2.60
